test: each file for each factory
