+++ protocol/synthetix/subgraph/tests/event-factories/index.ts
@@ -1,0 +1,25 @@
+export * from './createDepositEvent'
+export * from './createAccountCreatedEvent'
+export * from './createCollateralConfiguredEvent'
+export * from './createDelegationUpdatedEvent'
+export * from './createLiquidationEvent'
+export * from './createMarketRegisteredEvent'
+export * from './createMarketUSDDepositedEvent'
+export * from './createMarketUSDWithdrawnEvent'
+export * from './createPermissionGrantedEvent'
+export * from './createPermissionRevokedEvent'
+export * from './createPoolConfigurationSetEvent'
+export * from './createPoolCreatedEvent'
+export * from './createWithdrawEvent'
+export * from './createVaultLiquidationEvent'
+export * from './createUSDMintedEvent'
+export * from './createUSDBurnedEvent'
+export * from './createRewardsDistributorRemovedEvent'
+export * from './createRewardsDistributedEvent'
+export * from './createRewardsClaimedEvent'
+export * from './createRewardDistributorRegisteredEvent'
+export * from './createPoolOwnershipRenouncedEvent'
+export * from './createPoolNominationRevokedEvent'
+export * from './createPoolNameUpdatedEvent'
+export * from './createPoolOwnerNominatedEvent'
+export * from './createPoolOwnershipAcceptedEvent'
