+++ protocol/synthetix/subgraph/tests/event-factories/utils.ts
@@ -1,0 +1,6 @@
+export function createBlock(timestamp: i64, blockNumber: i64): Map<string, i64> {
+  const newBlock = new Map<string, i64>();
+  newBlock.set('timestamp', timestamp);
+  newBlock.set('blockNumber', blockNumber);
+  return newBlock;
+}
